fix(*.c files): Change long int to unsigned int in hexa, octal and binary functions and update main.h file with the same changes

+++ HEX_handler.c
@@ -6,7 +6,7 @@
  */
 int handler_HEX(va_list arg)
 {
-	long int num = va_arg(arg, long int);
+	unsigned int num = va_arg(arg, unsigned int);
 	char *str = to_HEX(num);
 	int strlen = _strlen(str);
